Modify vintage constraints
