+++ .jhipster/Vintage.json
@@ -22,7 +22,6 @@
             "fieldType": "Integer"
         }
     ],
-    "changelogDate": "20170303215710",
     "entityTableName": "vintage",
     "dto": "no",
     "pagination": "no",
