+++ src/main/resources/config/liquibase/changelog/20170303215710_added_entity_Cellar.xml
@@ -25,10 +25,10 @@
                 <constraints nullable="true" />
             </column>
             <column name="user_id" type="bigint">
-                <constraints unique="true" nullable="true" />
+                <constraints unique="true" nullable="false" />
             </column>
             <!-- jhipster-needle-liquibase-add-column - JHipster will add columns here, do not remove-->
         </createTable>
-        
+
     </changeSet>
 </databaseChangeLog>
