+++ src/main/resources/config/liquibase/changelog/20170805201621_added_entity_constraints_Vintage.xml
@@ -2,11 +2,11 @@
 <databaseChangeLog
     xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.4.xsd">
+    xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.5.xsd">
     <!--
         Added the constraints for entity Vintage.
     -->
-    <changeSet id="20170303215710-2" author="jhipster">
+    <changeSet id="20170805201621-2" author="jhipster">
         
         <addForeignKeyConstraint baseColumnNames="year_id"
                                  baseTableName="vintage"
