+++ src/main/webapp/app/entities/vintage/vintage-dialog.html
@@ -1,5 +1,5 @@
 
-<form name="editForm" role="form" novalidate ng-submit="vm.save()">
+<form name="editForm" role="form" novalidate ng-submit="vm.save()" show-validation>
 
     <div class="modal-header">
         <button type="button" class="close" data-dismiss="modal" aria-hidden="true"
@@ -22,15 +22,25 @@
 
         <div class="form-group">
             <label data-translate="cavavinApp.vintage.year" for="field_year">Year</label>
-            <select class="form-control" id="field_year" name="year" ng-model="vm.vintage.year" ng-options="year as year.number for year in vm.years track by year.id">
-                <option value=""></option>
+            <select class="form-control" id="field_year" name="year" ng-model="vm.vintage.year" ng-options="year as year.number for year in vm.years track by year.id" required>
             </select>
+        </div>
+        <div ng-show="editForm.year.$invalid">
+            <p class="help-block"
+                ng-show="editForm.year.$error.required" data-translate="entity.validation.required">
+                This field is required.
+            </p>
         </div>
         <div class="form-group">
             <label data-translate="cavavinApp.vintage.wine" for="field_wine">Wine</label>
-            <select class="form-control" id="field_wine" name="wine" ng-model="vm.vintage.wine" ng-options="wine as wine.name for wine in vm.wines track by wine.id">
-                <option value=""></option>
+            <select class="form-control" id="field_wine" name="wine" ng-model="vm.vintage.wine" ng-options="wine as wine.name for wine in vm.wines track by wine.id" required>
             </select>
+        </div>
+        <div ng-show="editForm.wine.$invalid">
+            <p class="help-block"
+                ng-show="editForm.wine.$error.required" data-translate="entity.validation.required">
+                This field is required.
+            </p>
         </div>
     </div>
     <div class="modal-footer">
