+++ src/test/gatling/user-files/simulations/VintageGatlingTest.scala
@@ -7,9 +7,9 @@ import org.slf4j.LoggerFactory
 import scala.concurrent.duration._
 
 /**
- * Performance test for the Wine entity.
+ * Performance test for the Vintage entity.
  */
-class WineGatlingTest extends Simulation {
+class VintageGatlingTest extends Simulation {
 
     val context: LoggerContext = LoggerFactory.getILoggerFactory.asInstanceOf[LoggerContext]
     // Log all HTTP requests
@@ -42,7 +42,7 @@ class WineGatlingTest extends Simulation {
         "Authorization" -> "${access_token}"
     )
 
-    val scn = scenario("Test the Wine entity")
+    val scn = scenario("Test the Vintage entity")
         .exec(http("First unauthenticated request")
         .get("/api/account")
         .headers(headers_http)
@@ -60,26 +60,26 @@ class WineGatlingTest extends Simulation {
         .check(status.is(200)))
         .pause(10)
         .repeat(2) {
-            exec(http("Get all wines")
-            .get("/api/wines")
+            exec(http("Get all vintages")
+            .get("/api/vintages")
             .headers(headers_http_authenticated)
             .check(status.is(200)))
             .pause(10 seconds, 20 seconds)
-            .exec(http("Create new wine")
-            .post("/api/wines")
+            .exec(http("Create new vintage")
+            .post("/api/vintages")
             .headers(headers_http_authenticated)
-            .body(StringBody("""{"id":null, "name":"SAMPLE_TEXT", "appellation":"SAMPLE_TEXT", "producer":"SAMPLE_TEXT", "creatorId":null}""")).asJSON
+            .body(StringBody("""{"id":null, "bareCode":"0"}""")).asJSON
             .check(status.is(201))
-            .check(headerRegex("Location", "(.*)").saveAs("new_wine_url"))).exitHereIfFailed
+            .check(headerRegex("Location", "(.*)").saveAs("new_vintage_url"))).exitHereIfFailed
             .pause(10)
             .repeat(5) {
-                exec(http("Get created wine")
-                .get("${new_wine_url}")
+                exec(http("Get created vintage")
+                .get("${new_vintage_url}")
                 .headers(headers_http_authenticated))
                 .pause(10)
             }
-            .exec(http("Delete created wine")
-            .delete("${new_wine_url}")
+            .exec(http("Delete created vintage")
+            .delete("${new_vintage_url}")
             .headers(headers_http_authenticated))
             .pause(10)
         }
@@ -87,6 +87,6 @@ class WineGatlingTest extends Simulation {
     val users = scenario("Users").exec(scn)
 
     setUp(
-        users.inject(rampUsers(100) over (1 minutes))
+        users.inject(rampUsers(Integer.getInteger("users", 100)) over (Integer.getInteger("ramp", 1) minutes))
     ).protocols(httpConf)
 }
